+++ packages/terra-form-input/src/InputField.jsx
@@ -76,6 +76,10 @@ const propTypes = {
    */
   onChange: PropTypes.func,
   /**
+   * Placeholder text.
+   */
+  placeholder: PropTypes.string,
+  /**
    * Ref callback to pass into the ref attribute of the html input element.
    */
   refCallback: PropTypes.func,
@@ -114,6 +118,7 @@ const defaultProps = {
   isLabelHidden: false,
   labelAttrs: {},
   onChange: undefined,
+  placeholder: undefined,
   maxWidth: undefined,
   refCallback: undefined,
   required: false,
@@ -140,6 +145,7 @@ const InputField = (props) => {
     labelAttrs,
     maxWidth,
     onChange,
+    placeholder,
     refCallback,
     required,
     showOptional,
@@ -163,6 +169,7 @@ const InputField = (props) => {
   }
 
   const inputType = type || inputAttrs.type;
+
   return (
     <Field
       label={label}
@@ -187,6 +194,7 @@ const InputField = (props) => {
         isIncomplete={isIncomplete}
         type={inputType}
         onChange={onChange}
+        placeholder={placeholder || inputAttrs.placeholder}
         value={value}
         defaultValue={defaultValue}
         refCallback={refCallback}