+++ packages/terra-form-input/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Added
 * Added `isIncomplete` prop to show incomplete stylings.
+* Added `placeholder` as a first class property to `<Input />` and `<InputField />`
 
 2.27.0 - (September 26, 2019)
 ------------------